style: `ResponseMaker` 목적의 패키지 생성
