+++ tomcat/src/main/java/org/apache/coyote/http11/responseGenerator/RegisterGetResponseMaker.java
@@ -1,4 +1,4 @@
-package org.apache.coyote.http11;
+package org.apache.coyote.http11.responseGenerator;
 
 import java.io.IOException;
 import java.net.URISyntaxException;
@@ -6,33 +6,22 @@ import java.net.URL;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.Paths;
-import java.util.Optional;
-import nextstep.jwp.model.User;
 import org.apache.coyote.http11.request.HttpRequest;
 import org.apache.coyote.http11.response.ContentType;
 import org.apache.coyote.http11.response.HttpResponse;
 import org.apache.coyote.http11.response.HttpStatus;
 
-public class LoginGetResponseMaker implements ResponseMaker {
+public class RegisterGetResponseMaker implements ResponseMaker {
 
     @Override
     public String createResponse(final HttpRequest httpRequest)
             throws URISyntaxException, IOException {
+        final String requestUrl = httpRequest.getRequestUrl();
         final URL resource =
-                this.getClass().getClassLoader().getResource("static" + "/login.html");
+                this.getClass().getClassLoader().getResource("static" + requestUrl + ".html");
         final Path path = Paths.get(resource.toURI());
         final var responseBody = new String(Files.readAllBytes(path));
-        final Optional<User> loginUser = httpRequest.findUserByJSessionId();
-        if (loginUser.isPresent()) {
-            return redirectByAlreadyLogin(responseBody);
-        }
         final HttpResponse httpResponse = new HttpResponse(HttpStatus.OK, responseBody, ContentType.HTML);
-        return httpResponse.toString();
-    }
-
-    private String redirectByAlreadyLogin(final String responseBody) {
-        final HttpResponse httpResponse =
-                new HttpResponse(HttpStatus.FOUND, responseBody, ContentType.HTML, "/index.html");
         return httpResponse.toString();
     }
 }
